feat: create Item class and move _decode* and _encode* methods
